MDEV-21117 Incremental commit fixes failover
The demoted into slave old master refused to accept own events by
existing rules. --replicate-same-server-id does not look to be a solution.

This is fixed with relaxing the same-server-id acceptence condition
for the semisync slave that connects to master in gtid mode.
This is safe to do when @@gtid_strict_mode = ON, which is recommended
therefore.